Move shared setUpClass to common module
Introduce a common.QiskitTestCase class that provides setUpClass(),
reducing the amount of duplicated code in the test cases.
Change permissions of some files that were set as executable.
