+++ test/python/test_qasm_python_simulator.py
@@ -27,23 +27,16 @@ if __name__ == '__main__':
 else:
     from test.python._random_qasm_generator import RandomQasmGenerator
 
-class LocalQasmSimulatorTest(unittest.TestCase):
+from .common import QiskitTestCase
+
+
+class LocalQasmSimulatorTest(QiskitTestCase):
     """Test local qasm simulator."""
 
     @classmethod
     def setUpClass(cls):
-        cls.moduleName = os.path.splitext(__file__)[0]
+        super(LocalQasmSimulatorTest, cls).setUpClass()
         cls.pdf = PdfPages(cls.moduleName + '.pdf')
-        cls.log = logging.getLogger(__name__)
-        cls.log.setLevel(logging.INFO)
-        logFileName = cls.moduleName + '.log'
-        handler = logging.FileHandler(logFileName)
-        handler.setLevel(logging.INFO)
-        log_fmt = ('{}.%(funcName)s:%(levelname)s:%(asctime)s:'
-                   ' %(message)s'.format(cls.__name__))
-        formatter = logging.Formatter(log_fmt)
-        handler.setFormatter(formatter)
-        cls.log.addHandler(handler)
 
     @classmethod
     def tearDownClass(cls):
